change column types

+++ migrations/sqls/20221031192845-order-table-up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE Orders(
     id SERIAL PRIMARY KEY,
-    product_id bigint NOT NULL,
+    product_id INTEGER NOT NULL,
     quantity INTEGER NOT NULL,
-    user_id bigint NOT NULL,
+    user_id INTEGER NOT NULL,
     status VARCHAR(20) NOT NULL
 );
 ALTER TABLE Orders
